feat: add restore_prod_db.py script, extract shared db_utils
Extract shared database utilities (parse_database_url, import_sql_dump,
create_test_user, etc.) from migrate_legacy_db.py into db_utils.py to
avoid duplication. Add restore_prod_db.py for importing prod dumps
without running legacy data migrations (BBCode, quotes, etc.).

Also: strip DEFINER clauses during SQL import to avoid privilege errors
when restoring prod dumps into local dev, fix group title casing for
test accounts, and add testtagger test user.
